Added advanced authentication features.

+++ Node-first-RESTAPI/controllers/feed.js
@@ -3,6 +3,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
 	//Stiamo usando dei parametri di tipo query, non parametri normali.
@@ -53,22 +54,30 @@ exports.createPost = (req, res, next) => {
 	const imageUrl = req.file.path;
 	const title = req.body.title;
 	const content = req.body.content;
+	let creator;
 
 	const post = new Post({
 		title: title,
 		content: content,
 		imageUrl: imageUrl,
-		creator: {
-			name: 'Riccardo',
-		},
+		creator: req.userId,
 	});
 	post.save()
 		.then((result) => {
-			console.log(result);
+			return User.findById(req.userId);
+		})
+		.then((user) => {
+			creator = user;
+			//Possiamo passare anche solo il post. Sarà poi mongoDB a capire qual'è l'id corretto.
+			user.posts.push(post);
+			return user.save();
+		})
+		.then((result) => {
 			//201 come status code è specifico nel dire che una risorsa è stata creata con successo.
 			res.status(201).json({
 				message: 'Post created succesfully!',
-				post: result,
+				post: post,
+				creator: { _id: creator._id, name: creator.name },
 			});
 		})
 		.catch((err) => {
@@ -133,6 +142,12 @@ exports.updatePost = (req, res, next) => {
 				throw error;
 			}
 
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized action.');
+				error.statusCode = 403;
+				throw error;
+			}
+
 			if (imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl);
 			}
@@ -164,10 +179,23 @@ exports.deletePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
+
 			//Check logged user
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized action.');
+				error.statusCode = 403;
+				throw error;
+			}
 
 			clearImage(post.imageUrl);
 			return Post.findByIdAndRemove(postId);
+		})
+		.then((result) => {
+			return User.findById(req.userId);
+		})
+		.then((user) => {
+			user.posts.pull(postId);
+			return user.save();
 		})
 		.then((result) => {
 			console.log(result);
